Move resources contract to own module
